Updates 0-print_listint.c file with %u

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -7,13 +7,13 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int count;
+	size_t count;
 
 	if (h == NULL)
 		return (0);
 	for (count = 0; h != NULL; count++)
 	{
-		printf("%d\n", h->n);
+		printf("%u\n", h->n);
 		h = h->next;
 	}
 	return (count);
